merge develop into main

+++ template/parser.py
@@ -1,10 +1,11 @@
-import sys
-sys.path.append("../")
-from tools import Parser, logfile
+from utils.logger import logfile, LogLevel
+from utils.box import Box
+from core.config import ConfigManager
+from core.parser import Parser
 from importlib import import_module
-import os
-import shutil
 import json
+import shutil
+import os
 
 class ParserTemplate(Parser):
 
@@ -12,41 +13,44 @@ class ParserTemplate(Parser):
         Parser.__init__(self)
         self.clearExistFiles()
 
-    def doParser(self):
+    def doParse(self):
         self.tidy()
-        # your parser script
-        writerList = os.listdir(".\\clazz")
-        for fileName in writerList:
-            if fileName == "__init__.py" or fileName[-3:] != ".py":
+        clazzPath = Box.get(ConfigManager).get("Clazz")
+        if not clazzPath:
+            logfile(LogLevel.ERROR, "no clazz path")
+            return
+        writerList = os.listdir(clazzPath)
+        for filename in writerList:
+            if filename == "__init__.py" or filename[-3:] != ".py":
                 continue            
-            target = import_module(f"project.clazz.{fileName[:-3]}")
+            target = import_module(f"clazz.{filename[:-3]}")
             if not target:
-                logfile("parser", "error writer file path")
+                logfile(LogLevel.ERROR, "error writer file path")
                 continue
             if not hasattr(target, "Writer"):
-                logfile("parser", f"file {fileName} no writer")
+                logfile(LogLevel.ERROR, f"file {filename} no writer")
                 continue
             Writer = getattr(target, "Writer")
-            Writer(self.conf).write()
+            Writer().write()
 
     def tidy(self):
-        path = ".\\data\middleware"
-        for fileName in os.listdir(path):
-            if fileName[-5:] != ".json":
+        path = Box.get(ConfigManager).get("MiddleFile")
+        for filename in os.listdir(path):
+            if filename[-5:] != ".json":
                 continue
-            with open(path + "\\" + fileName, "r", encoding="utf-8") as file1:
+            with open(os.path.join(path, filename), "r", encoding="utf-8") as file1:
                 jsonObj = json.loads(file1.read())
                 jsonObj = sorted(jsonObj.items(), key = lambda x:int(x[0]) if isinstance(x[0], int) else x[0])
                 jsonObj = dict(jsonObj)
-                with open(path + "\\" + fileName, "w", encoding="utf-8") as file2:
+                with open(os.path.join(path, filename), "w", encoding="utf-8") as file2:
                     file2.write(json.dumps(jsonObj, indent=4, ensure_ascii=False))
 
     def clearExistFiles(self):
-        path = ".\\data\middleware"
-        shutil.rmtree(path)
+        path = Box.get(ConfigManager).get("MiddleFile")
+        if os.path.exists(path):
+            shutil.rmtree(path)
         os.mkdir(path)
 
 
 if __name__ == "__main__":
-    parser = ParserTemplate()
-    parser.parser()
+    ParserTemplate().parse()
